[xs] change name of authorization method is_allowed
The name of the main function in authorization was is_allowed. Although the name
holds right meaning in the scope of this file but in global scope the name was bit
confusing. The ambiguity was, what is allowed. That is why changing the name
of the method to is_authorize. The new name will not be confusing in global context.

Ref - #260