[xs] fix navbar search box
The search box in navbar was not responding to enter button.
The main reason was the javascript search method that attached to search button.

The form input was required mode. So blank search was not working.

Removed the search button, search function in js and required mode.

Another bug found was while clicking main-logo it supposed to redirect to home page.
But it was redirecting to '#'. Updated the href to '/'

Ref - #261