[m] change SQL-Alchemy first call to one
Previously in Package model we had publisher name, package name
and version as an unique key. That is why filter call with publisher
name and package name was returning multiple rows.

But now in Package model publisher name and package name is unique
key. So there must be only one row for filtering publisher name and
package name. That is why first method call is ambiguous.
Changing to one method call.

Resolves : #313